Make it a package
