+++ c_api/python/taichi_runtime/__init__.py
@@ -1,0 +1,4 @@
+from _lib import _LIB
+from taichi_core import *
+from taichi_metal import *
+from taichi_vulkan import *
